good night : )

+++ src/SharedObject/IRenderable.java
@@ -3,7 +3,7 @@ package SharedObject;
 import javafx.scene.canvas.GraphicsContext;
 
 public interface IRenderable {
-	public int getZ();
+	public double getZ();
 	public void draw(GraphicsContext gc);
 	public boolean isDestroyed();
 	public boolean isVisible();
